// src/app/api/models/address-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AddressDto {
  addressLine1: string;
  addressLine2?: string;
  city: string;
  country: string;
  createdAt?: string;
  customerEmail?: string;
  formattedAddress?: string;
  fullName: string;
  id?: number;
  isDefault?: boolean;
  notes?: string;
  phoneNumber: string;
  postalCode: string;
  state: string;
  type?: string;
  updatedAt?: string;
}